Export RunE2EKit so other kit repos can reuse the real-sandbox e2e layer
TestE2EKit's logic lived entirely inside a _test.go file, which Go never
compiles into what an importing module gets — so nothing outside this repo
could reuse the real-sandbox assertions the way tck.NewSuiteFromDir/RunAll
already let importers reuse the container-based TCK layer. Same problem for
secretServiceStored, used by the newer checkHostCredentials skip-check: it
lived in tck/hostcredentials_test.go, also unreachable from outside.

Move that logic into an exported tck.RunE2EKit(t, kitPath, E2EOptions) in an
ordinary tck/e2e.go, parameterized by an app-name (previously hardcoded to
this repo's own "sbx-kits-contrib-tck") so a caller can scope every `sbx`
invocation to its own daemon. tck/e2e_test.go's TestE2EKit becomes a thin
wrapper — same behavior, same env var contract. secretServiceStored moves
into e2e.go too; its unit test (tck/hostcredentials_test.go) becomes a
white-box test in package tck calling it directly.

Rebased onto main to pick up the host-credential skip check and kit-args
substitution added since this PR was opened; verified against a live kit
via ./scripts/test-kit.sh to confirm the re-applied refactor didn't change
behavior.
