+++ tck/e2e_test.go
@@ -5,634 +5,33 @@
 // runs. Build-tagged `e2e` so it never runs in the default `go test ./...`
 // flow that kit authors invoke locally — only the matrix job in tck.yml
 // opts in via `-tags=e2e`.
+//
+// This is a thin wrapper around the exported tck.RunE2EKit — see tck/e2e.go
+// for the actual e2e logic, which any module importing this package (e.g.
+// sbx-kits-internal) can drive against its own app-name.
 
 package tck_test
 
 import (
-	"context"
-	"crypto/rand"
-	"encoding/hex"
-	"errors"
-	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 	"testing"
-	"time"
 
-	"github.com/docker/sbx-kits-contrib/spec"
 	"github.com/docker/sbx-kits-contrib/tck"
 	"github.com/stretchr/testify/require"
-	"go.yaml.in/yaml/v3"
 )
 
-// appName scopes every sbx call in this suite to a dedicated daemon instance,
-// isolated from the operator's default sbx state.
+// appName scopes every `sbx` call this repo's own e2e run makes, keeping it
+// isolated from a developer's or another module's main sbx state. Must stay
+// in sync with scripts/test-kit-e2e.sh's default APP_NAME.
 const appName = "sbx-kits-contrib-tck"
 
-// sandboxWorkDir is the workspace mount point inside every sbx template
-// container. All templates inherit from the shared `base` stage which sets
-// `WORKDIR /home/agent/workspace`, so `${WORKDIR}` placeholders in a kit's
-// `commands.initFiles` resolve to this path at runtime. (The tck.TestWorkDir
-// constant is `/workspace` because the in-suite testcontainers tests fabricate
-// their own workdir; it does not match a real sbx sandbox.)
-const sandboxWorkDir = "/home/agent/workspace"
-
-// TestE2EKit is the single e2e entry point for all kit types. It creates one
-// sandbox, verifies kit content (env vars, files, tmpfs, agentContext) for
-// every kit, and — for kind:sandbox kits that declare promptArgs in their
-// testdata/tck.yaml — also sends a non-interactive prompt to the agent.
-//
-// Subtest layout:
-//
-//	TestE2EKit/<kit>
-//	  env          — all kits
-//	  files        — all kits
-//	  tmpfs        — all kits
-//	  agentContext — all kits (skipped when agentContext is absent)
-//	  prompt       — kind:sandbox only, skipped when tck.yaml/promptArgs absent
-//
-// The agent passed to `sbx create` depends on the kit's manifest:
-//
-//	kind: sandbox → the kit's own name (sbx enforces this match).
-//	kind: mixin   → "claude" (the default agent the kit is exercised against).
-//
-// KIT_UNDER_TEST is read from the environment; the CI matrix sets it per-kit.
+// TestE2EKit is the single e2e entry point for all kit types in this repo.
+// KIT_UNDER_TEST is read from the environment; the CI matrix and
+// scripts/test-kit-e2e.sh set it per-kit. See tck.RunE2EKit for what it
+// actually does (subtests: env, files, tmpfs, agentContext, prompt).
 func TestE2EKit(t *testing.T) {
 	kitPath := os.Getenv("KIT_UNDER_TEST")
 	require.NotEmpty(t, kitPath, "KIT_UNDER_TEST must point at a kit directory")
 
-	absKit, err := filepath.Abs(kitPath)
-	require.NoError(t, err, "resolve KIT_UNDER_TEST=%q", kitPath)
-
-	t.Run(filepath.Base(absKit), func(t *testing.T) {
-		info, err := os.Stat(absKit)
-		require.NoErrorf(t, err, "stat KIT_UNDER_TEST=%q", absKit)
-		require.Truef(t, info.IsDir(), "KIT_UNDER_TEST=%q must be a directory", absKit)
-
-		suite, err := tck.NewSuiteFromDir(absKit)
-		require.NoErrorf(t, err, "derive suite for %q", absKit)
-
-		_, err = exec.LookPath("sbx")
-		require.NoError(t, err, "sbx must be on PATH; CI installs it from docker/sbx-releases")
-
-		agent := agentForKit(suite.Artifact)
-
-		// Loaded before `sbx create` rather than just before the prompt
-		// subtest: it may declare that a built-in agent still shadows this
-		// kit's name, which changes how a create failure is interpreted.
-		td := loadKitTCKData(t, absKit)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Minute)
-		defer cancel()
-
-		checkHostCredentials(t, ctx, td)
-
-		name := createSbx(t, ctx, absKit, suite.Artifact.Manifest.Name, agent, td)
-
-		// Verify kit content landed inside the running container. Files are
-		// re-derived here so `${WORKDIR}` resolves to the real sandbox workdir
-		// instead of the TCK in-suite constant. These run for all kit kinds.
-		assertSbxEnv(t, ctx, name, suite.ExpectedEnvVars)
-		assertSbxFiles(t, ctx, name, expectedSandboxFiles(suite.Artifact))
-		assertSbxTmpfs(t, ctx, name, suite.ExpectedTmpfs)
-		assertSbxAgentContext(t, ctx, name, suite.Artifact)
-
-		// Prompt subtest: kind:sandbox only, opt-in via testdata/tck.yaml.
-		if suite.Artifact.Manifest.Kind != spec.KindSandbox {
-			return
-		}
-		if td == nil || len(td.PromptArgs) == 0 {
-			return
-		}
-
-		// Wait for any background installation before sending the prompt.
-		waitForAgentReady(t, ctx, name, td.ReadyFile)
-
-		binary := td.Binary
-		if binary == "" {
-			binary = filepath.Base(suite.Artifact.Manifest.Binary)
-		}
-		require.NotEmptyf(t, binary,
-			"kit %s: binary not derivable from spec.yaml entrypoint or tck.yaml", suite.Artifact.Manifest.Name)
-
-		t.Run("prompt", func(t *testing.T) {
-			execArgs := []string{"exec", name, "--", binary}
-			execArgs = append(execArgs, td.PromptArgs...)
-			execArgs = append(execArgs, promptMessage)
-
-			// Ignore exit error: auth failures (401/403) are expected without
-			// credentials. The assertion is that the agent ran and responded.
-			out, _ := runSbx(t, ctx, execArgs...)
-			require.NotEmptyf(t, out,
-				"sbx exec produced no output for kit %s (binary=%s, promptArgs=%v)",
-				absKit, binary, td.PromptArgs)
-			t.Logf("sbx exec output for kit %s (binary=%s, promptArgs=%v):\n%s",
-				absKit, binary, td.PromptArgs, out)
-		})
-	})
-}
-
-// expectedSandboxFiles derives the absolute paths a kit must materialize
-// inside a real sbx sandbox: every file under `files/home` lands at
-// /home/agent/<relpath>, every `commands.initFiles` entry lands at its
-// declared path with `${WORKDIR}` replaced by the actual sandbox workdir.
-func expectedSandboxFiles(a *spec.Artifact) []string {
-	var paths []string
-	for _, f := range a.Files {
-		if f.Target == spec.TargetHome {
-			paths = append(paths, tck.HomeDir+"/"+f.RelativePath)
-		}
-	}
-	if a.Commands != nil {
-		for _, f := range a.Commands.InitFiles {
-			paths = append(paths, strings.ReplaceAll(f.Path, "${WORKDIR}", sandboxWorkDir))
-		}
-	}
-	return paths
-}
-
-// assertSbxEnv runs `printenv` in the sandbox and checks each declared
-// environment variable from `environment.variables` is set to the expected
-// value.
-func assertSbxEnv(t *testing.T, ctx context.Context, name string, expected []string) {
-	if len(expected) == 0 {
-		return
-	}
-	t.Run("env", func(t *testing.T) {
-		out, err := runSbx(t, ctx, "exec", name, "--", "env")
-		require.NoErrorf(t, err, "sbx exec env failed:\n%s", out)
-		for _, kv := range expected {
-			require.Containsf(t, out, kv,
-				"sandbox env should contain %q\nfull env output:\n%s", kv, out)
-		}
-	})
-}
-
-// assertSbxFiles checks that each expected file exists and is non-empty inside
-// the sandbox.
-func assertSbxFiles(t *testing.T, ctx context.Context, name string, expected []string) {
-	if len(expected) == 0 {
-		return
-	}
-	t.Run("files", func(t *testing.T) {
-		for _, path := range expected {
-			path := path
-			t.Run(path, func(t *testing.T) {
-				out, err := runSbx(t, ctx, "exec", name, "--", "test", "-s", path)
-				require.NoErrorf(t, err,
-					"file %q should exist and be non-empty in sandbox:\n%s", path, out)
-			})
-		}
-	})
-}
-
-// assertSbxTmpfs verifies each expected tmpfs path is mounted as tmpfs inside
-// the sandbox. The /run/secrets entry is always present and is checked along
-// with any manifest-declared tmpfs paths.
-func assertSbxTmpfs(t *testing.T, ctx context.Context, name string, expected map[string]string) {
-	if len(expected) == 0 {
-		return
-	}
-	t.Run("tmpfs", func(t *testing.T) {
-		out, err := runSbx(t, ctx, "exec", name, "--", "mount")
-		require.NoErrorf(t, err, "sbx exec mount failed:\n%s", out)
-		for path := range expected {
-			path := path
-			t.Run(path, func(t *testing.T) {
-				marker := fmt.Sprintf("tmpfs on %s ", path)
-				require.Containsf(t, out, marker,
-					"%s should be mounted as tmpfs; mount output:\n%s", path, out)
-			})
-		}
-	})
-}
-
-// assertSbxAgentContext verifies that a kit's `agentContext:` content was
-// rendered into a file inside the sandbox. For `kind: sandbox` kits the
-// agentContext is inlined in the AI profile file (Manifest.AIFilename); for
-// `kind: mixin` kits the engine writes a per-kit file as
-// `kits-agent-context/<kit-name>.md` next to the parent agent's AI file. The
-// exact directory varies per template, so we locate candidate files by name
-// first, then grep each one for a stable substring of the declared content.
-func assertSbxAgentContext(t *testing.T, ctx context.Context, name string, a *spec.Artifact) {
-	if a.AgentContext == "" {
-		return
-	}
-	target := agentContextFilename(a)
-	if target == "" {
-		t.Logf("agentContext check skipped: kit %s declares agentContext but has no aiFilename (kind=%s)",
-			a.Manifest.Name, a.Manifest.Kind)
-		return
-	}
-	needle := agentContextNeedle(a.AgentContext)
-	if needle == "" {
-		t.Logf("agentContext check skipped: no usable line in declared agentContext")
-		return
-	}
-	t.Run("agentContext", func(t *testing.T) {
-		// Find candidates by filename in writable trees. Prune pseudo-fs to
-		// keep the traversal fast and avoid permission noise. `|| true`
-		// hides find's non-zero exit when it bumps into unreadable
-		// subdirectories — we judge by the printed paths, not the status.
-		findCmd := fmt.Sprintf(
-			"find / -path /proc -prune -o -path /sys -prune -o -path /dev -prune -o "+
-				"-type f -name %s -print 2>/dev/null || true",
-			shellQuote(target),
-		)
-		findOut, err := runSbx(t, ctx, "exec", name, "--", "sh", "-c", findCmd)
-		require.NoErrorf(t, err, "sbx exec find failed:\n%s", findOut)
-
-		paths := strings.Fields(findOut)
-		require.NotEmptyf(t, paths,
-			"no agentContext file %q found anywhere in sandbox (kind=%s, name=%s)",
-			target, a.Manifest.Kind, a.Manifest.Name)
-
-		for _, p := range paths {
-			grepCmd := fmt.Sprintf("grep -qF -- %s %s", shellQuote(needle), shellQuote(p))
-			if _, err := runSbx(t, ctx, "exec", name, "--", "sh", "-c", grepCmd); err == nil {
-				t.Logf("agentContext matched in %s", p)
-				return
-			}
-		}
-		t.Fatalf("agentContext needle %q not found in any candidate %v (kind=%s, name=%s)",
-			needle, paths, a.Manifest.Kind, a.Manifest.Name)
-	})
-}
-
-// agentContextFilename returns the filename the engine writes a kit's
-// agentContext into.
-//
-//	kind: sandbox → Manifest.AIFilename (inlined agentContext)
-//	kind: mixin   → "<kit-name>.md" under .../kits-agent-context/
-func agentContextFilename(a *spec.Artifact) string {
-	if a.Manifest.Kind == spec.KindSandbox {
-		return a.Manifest.AIFilename
-	}
-	if a.Manifest.Kind == spec.KindMixin {
-		return a.Manifest.Name + ".md"
-	}
-	return ""
-}
-
-// agentContextNeedle picks the longest non-empty, non-fenced, no-backtick
-// line from the declared agentContext to use as a grep target. Avoiding
-// backticks keeps the quoting story simple; picking the longest line
-// minimizes the chance the substring collides with boilerplate.
-func agentContextNeedle(agentContext string) string {
-	var best string
-	for _, raw := range strings.Split(agentContext, "\n") {
-		line := strings.TrimSpace(raw)
-		if line == "" || strings.HasPrefix(line, "```") || strings.ContainsAny(line, "`'") {
-			continue
-		}
-		if len(line) > len(best) {
-			best = line
-		}
-	}
-	if len(best) > 80 {
-		best = best[:80]
-	}
-	return best
-}
-
-// kitTCKData holds kit-specific configuration for TCK e2e tests, loaded from
-// <kit-dir>/testdata/tck.yaml.
-type kitTCKData struct {
-	// ReadyFile is the absolute path of a sentinel file written inside the
-	// sandbox when a background installation completes. When set,
-	// TestE2ERunAgent polls `sbx exec -- test -f <readyFile>` before running
-	// the prompt subtest. Leave empty for kits whose install commands are synchronous.
-	ReadyFile string `yaml:"readyFile"`
-
-	// PromptArgs are the arguments prepended before the prompt message when
-	// invoking the binary non-interactively via `sbx exec`, e.g. ["-p"] for
-	// claude, ["-m"] for nanobot, ["chat", "-q"] for hermes. An empty slice
-	// skips the prompt subtest (for agents with no non-interactive mode).
-	PromptArgs []string `yaml:"promptArgs"`
-
-	// Binary overrides the agent binary name used in `sbx exec`. When absent,
-	// filepath.Base(suite.Artifact.Manifest.Binary) is used — the normalizer
-	// derives that from sandbox.entrypoint.run[0]. Set this only when the
-	// entrypoint is a wrapper script whose real binary has a different name
-	// (e.g. nanoclaw's nanoclaw-start wraps claude).
-	Binary string `yaml:"binary"`
-
-	// ExtractedFromBuiltin marks a kind:sandbox kit whose name is (still) that
-	// of an agent built into sbx. sbx refuses to load such a kit, failing with
-	// "already registered (built-in agents cannot be overridden by a kit)", so
-	// `sbx create` cannot succeed until a released sbx drops the built-in.
-	//
-	// Setting this to true turns that specific failure into a SKIP instead of a
-	// failure, which lets an extracted kit land and be reviewed in this repo
-	// before the sbx side of the extraction ships. Every other failure mode
-	// still fails.
-	//
-	// It is deliberately narrow, not a blanket "skip e2e":
-	//
-	//   * The skip only triggers on the collision error itself. Once sbx no
-	//     longer registers the built-in, `sbx create` succeeds and the full e2e
-	//     runs — nothing to remember to re-enable.
-	//   * Without this flag the collision is a hard failure, so a kit that
-	//     accidentally takes a built-in's name (a contributor naming a kit
-	//     "claude") is still caught.
-	//   * When the flag is set but creation succeeds, the test logs a notice
-	//     that the flag is now obsolete and should be deleted.
-	ExtractedFromBuiltin bool `yaml:"extractedFromBuiltin"`
-
-	// RequiresHostCredentials names `sbx secret` service(s) that must already
-	// be stored on the host — via `sbx --app-name sbx-kits-contrib-tck secret
-	// set <service>` — before this kit's agent can be created at all. Every
-	// listed service is checked before `sbx create` runs; if any is missing,
-	// the whole kit e2e is skipped naming the missing service(s) and the
-	// command to store them.
-	//
-	// Unlike ExtractedFromBuiltin, this does not self-obsolete: the agent
-	// keeps needing the credential indefinitely, so the entry is permanent
-	// for as long as the kit declares this agent affinity.
-	RequiresHostCredentials []string `yaml:"requiresHostCredentials"`
-
-	// Args are the values to substitute for the arguments the kit declares,
-	// keyed by argument name. The TCK layer reads the same block, so both
-	// layers install the kit with the same values: here they are passed to
-	// `sbx create` as kit-argument flags. An argument absent from this block
-	// resolves to its declared default on both sides and needs no entry.
-	Args map[string]string `yaml:"args"`
-}
-
-// builtinCollisionMarker is the distinguishing text of the error sbx returns
-// when a kit's name collides with a built-in agent. Matched as a substring
-// because the message embeds the agent name:
-//
-//	ERROR: agent "kiro" is already registered (built-in agents cannot be overridden by a kit)
-//
-// This couples to sbx's user-facing error text, so it can break if that wording
-// changes. The failure mode is safe — a reworded message stops matching and the
-// collision becomes an ordinary test failure rather than a silent skip.
-const builtinCollisionMarker = "built-in agents cannot be overridden by a kit"
-
-// loadKitTCKData reads <kitDir>/testdata/tck.yaml. Returns nil, nil when the
-// file does not exist — callers should skip rather than fail in that case.
-func loadKitTCKData(t *testing.T, kitDir string) *kitTCKData {
-	t.Helper()
-
-	p := filepath.Join(kitDir, "testdata", "tck.yaml")
-	data, err := os.ReadFile(p)
-	if os.IsNotExist(err) {
-		return nil
-	}
-	require.NoErrorf(t, err, "read %s", p)
-
-	var td kitTCKData
-	if err := yaml.Unmarshal(data, &td); err != nil {
-		require.NoErrorf(t, err, "parse %s", p)
-	}
-	return &td
-}
-
-// checkHostCredentials skips the kit's e2e run if any service in
-// td.RequiresHostCredentials has no secret stored on this sbx daemon. It
-// probes for presence rather than matching `sbx create`'s failure text: that
-// failure gives no indication of a missing credential, so string-matching it
-// would be fragile and, being generic, could mask unrelated create failures
-// as a skip. A probe failure or an unparseable probe result therefore fails
-// the test instead of skipping — only a clean "stored count is zero" does.
-// No-op when td is nil or declares no required services.
-func checkHostCredentials(t *testing.T, ctx context.Context, td *kitTCKData) {
-	t.Helper()
-	if td == nil || len(td.RequiresHostCredentials) == 0 {
-		return
-	}
-
-	var missing []string
-	for _, svc := range td.RequiresHostCredentials {
-		out, err := runSbxStdout(t, ctx, "secret", "ls", "--service", svc, "--json")
-		if err != nil {
-			var exitErr *exec.ExitError
-			if errors.As(err, &exitErr) {
-				out += string(exitErr.Stderr)
-			}
-		}
-		require.NoErrorf(t, err, "probe for a stored %q secret failed:\n%s", svc, out)
-
-		stored, perr := secretServiceStored([]byte(out))
-		require.NoErrorf(t, perr, "parse `sbx secret ls --service %s --json` output:\n%s", svc, out)
-
-		if !stored {
-			missing = append(missing, svc)
-		}
-	}
-	if len(missing) == 0 {
-		return
-	}
-
-	var setCmds strings.Builder
-	for _, svc := range missing {
-		fmt.Fprintf(&setCmds, "  sbx --app-name %s secret set %s\n", appName, svc)
-	}
-	t.Skipf("kit's agent needs credential(s) not stored on this sbx daemon: %s.\n"+
-		"This is expected in CI, which stores no per-agent cloud credentials — the requirement "+
-		"is permanent for this agent, not something to remove. To run the full e2e locally, "+
-		"store the missing service(s) first:\n%s",
-		strings.Join(missing, ", "), setCmds.String())
-}
-
-// promptMessage is the fixed text sent to every agent in TestE2EKit's prompt subtest.
-// It is short, benign, and produces a non-empty response even when the
-// agent rejects the call with an authentication error.
-const promptMessage = "what version are you running"
-
-// createSbx creates a sandbox for the kit using `sbx create`, registers a
-// t.Cleanup that force-removes it, and returns the sandbox name. A fresh temp
-// dir is used as the workspace.
-// td may be nil (no testdata/tck.yaml); see kitTCKData.ExtractedFromBuiltin for
-// the one failure this tolerates.
-func createSbx(t *testing.T, ctx context.Context, absKit, kitName, agent string, td *kitTCKData) string {
-	t.Helper()
-	workspace := t.TempDir()
-	name := sandboxName(t, absKit)
-
-	// Record the name for scripts/test-kit-e2e.sh's on_exit trap, regardless
-	// of what happens next: `sbx create` itself tears the sandbox down on a
-	// failed kit-apply/container-run (sandboxd's Provision/Run rollback) —
-	// the most common e2e failure — so by the time this test process exits,
-	// `sbx ls` may show nothing to find at all. The daemon's network/policy
-	// log is independent of the sandbox object's lifetime (it's a daemon-
-	// level log filtered by VM name, not per-sandbox state), so `sbx policy
-	// log <name>` still works after that rollback as long as the name is
-	// known — this file is the only way the wrapping shell script can learn
-	// a name generated randomly inside this process. Best-effort: a failure
-	// to record it must not fail the test itself.
-	if logPath := os.Getenv("SBX_E2E_NAME_LOG"); logPath != "" {
-		if f, err := os.OpenFile(logPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644); err == nil {
-			_, _ = fmt.Fprintln(f, name)
-			_ = f.Close()
-		}
-	}
-
-	t.Cleanup(func() {
-		// Leave a failed test's sandbox behind instead of removing it: on
-		// any failure that does NOT already trigger sandboxd's own rollback
-		// (e.g. an assertion failing after a successful create), this is
-		// the only post-mortem evidence of what went wrong. t.Cleanup runs
-		// during the normal failure unwind of require.NoErrorf/t.Fatal,
-		// before the process ever exits.
-		if t.Failed() {
-			t.Logf("keeping sandbox %q for post-mortem: sbx --app-name %s policy log %s", name, appName, name)
-			return
-		}
-		cleanCtx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		defer cancel()
-		out, err := exec.CommandContext(cleanCtx, "sbx", "--app-name", appName, "rm", "-f", name).CombinedOutput()
-		// A sandbox that was never created has nothing to remove. Reporting
-		// that as a cleanup failure buries the real error under a misleading
-		// "not found" whenever `sbx create` itself failed.
-		if err != nil && !strings.Contains(string(out), "not found") {
-			t.Logf("cleanup `sbx rm -f %s` failed: %v\n%s", name, err, out)
-		}
-	})
-	createArgs := []string{"create", "--kit", absKit, "--name", name}
-	if td != nil {
-		createArgs = append(createArgs, tck.KitArgFlags(kitName, td.Args)...)
-	}
-	createArgs = append(createArgs, agent, workspace)
-	createOut, err := runSbx(t, ctx, createArgs...)
-
-	if err != nil && strings.Contains(createOut, builtinCollisionMarker) {
-		if td != nil && td.ExtractedFromBuiltin {
-			t.Skipf("kit %q is still shadowed by the built-in %q agent in this sbx build, and "+
-				"declares extractedFromBuiltin: true — skipping e2e.\n"+
-				"This resolves itself once a released sbx drops the built-in; no action needed here.\n%s",
-				filepath.Base(absKit), agent, createOut)
-		}
-		// Same error without the declaration: almost always a kit that has
-		// taken a built-in agent's name by mistake. Say so explicitly, since
-		// the raw sbx message does not mention the two ways out.
-		require.FailNowf(t, "kit name collides with a built-in agent",
-			"kit %q cannot be loaded because sbx already registers %q as a built-in agent.\n"+
-				"Either rename the kit, or — if this kit is intentionally replacing that built-in — "+
-				"set `extractedFromBuiltin: true` in %s/testdata/tck.yaml.\n%s",
-			filepath.Base(absKit), agent, filepath.Base(absKit), createOut)
-	}
-
-	require.NoErrorf(t, err, "sbx create failed (agent=%s):\n%s", agent, createOut)
-
-	// The kit declared a built-in collision but none happened. Report the
-	// observation without drawing the conclusion: "no collision" has two very
-	// different causes, and only one of them means the flag is obsolete.
-	//
-	//   1. A released sbx dropped the built-in — the flag has done its job and
-	//      should be deleted.
-	//   2. The kit is being run under a different name (an author renaming it
-	//      locally to get past the collision and exercise the real create path).
-	//      The flag is still needed and deleting it would be wrong.
-	//
-	// Nothing here can tell those apart, so an unconditional "remove the flag"
-	// would hand out bad advice in exactly the case authors hit while extracting
-	// a built-in agent.
-	if td != nil && td.ExtractedFromBuiltin {
-		t.Logf("NOTICE: kit %q declares `extractedFromBuiltin: true`, but creating it as "+
-			"agent %q hit no built-in collision. If a released sbx has dropped the built-in, "+
-			"the flag is obsolete and can be removed from testdata/tck.yaml. If this ran under "+
-			"a renamed agent, the flag is still required — leave it.",
-			filepath.Base(absKit), agent)
-	}
-
-	t.Logf("sbx create succeeded for kit %s as sandbox %q (agent=%s)\n%s", absKit, name, agent, createOut)
-	return name
-}
-
-// waitForAgentReady polls until readyFile exists inside the sandbox, indicating
-// that a background installation completed. No-op when readyFile is empty
-// (synchronous-install kits are ready as soon as sbx create returns).
-func waitForAgentReady(t *testing.T, ctx context.Context, name, readyFile string) {
-	t.Helper()
-	if readyFile == "" {
-		return
-	}
-	t.Logf("polling for %s in sandbox %s...", readyFile, name)
-	for {
-		_, err := runSbx(t, ctx, "exec", name, "--", "test", "-f", readyFile)
-		if err == nil {
-			t.Logf("%s present — agent installation complete", readyFile)
-			return
-		}
-		select {
-		case <-ctx.Done():
-			t.Fatalf("timed out waiting for %s in sandbox %s — installation did not complete", readyFile, name)
-			return
-		case <-time.After(10 * time.Second):
-		}
-	}
-}
-
-// agentForKit picks the positional agent argument for `sbx create`. Sandbox
-// kits must be invoked with their own name as the agent. A mixin that declares
-// base-agent affinity (requires.agent) must be exercised on that agent —
-// composing it onto any other base is a hard error, since affinity is enforced
-// at compose time. Mixins without affinity default to claude.
-func agentForKit(a *spec.Artifact) string {
-	if a.Manifest.Kind == spec.KindSandbox {
-		return a.Manifest.Name
-	}
-	if a.Requires != nil && a.Requires.Agent != "" {
-		return a.Requires.Agent
-	}
-	return "claude"
-}
-
-// sandboxName builds a unique, sbx-name-safe identifier from the kit
-// directory and a random suffix. sbx accepts letters, numbers, hyphens,
-// periods, plus and minus signs — no underscores.
-func sandboxName(t *testing.T, kitDir string) string {
-	t.Helper()
-
-	base := strings.ToLower(filepath.Base(kitDir))
-	base = strings.ReplaceAll(base, "_", "-")
-
-	var raw [4]byte
-	_, err := rand.Read(raw[:])
-	require.NoError(t, err)
-
-	return "e2e-" + base + "-" + hex.EncodeToString(raw[:])
-}
-
-func buildSbxCmd(ctx context.Context, args ...string) *exec.Cmd {
-	all := make([]string, 0, len(args)+2)
-	all = append(all, "--app-name", appName)
-	all = append(all, args...)
-	cmd := exec.CommandContext(ctx, "sbx", all...)
-	cmd.Env = os.Environ()
-	return cmd
-}
-
-// runSbx invokes the sbx CLI and returns combined stdout+stderr. Inherits the
-// current environment so secrets and credential stores set up by the CI
-// `sbx login` step flow through. --app-name is prepended to every call for
-// telemetry attribution. Marked as a test helper so require/Fatal failures
-// inside callers point at the call site, not this wrapper.
-func runSbx(t *testing.T, ctx context.Context, args ...string) (string, error) {
-	t.Helper()
-	cmd := buildSbxCmd(ctx, args...)
-	out, err := cmd.CombinedOutput()
-	return string(out), err
-}
-
-// runSbxStdout invokes the sbx CLI like runSbx but returns stdout only,
-// discarding stderr. Callers that decode --json output need this: any stderr
-// text sbx writes alongside a JSON payload would otherwise be interleaved
-// into it and break decoding. On a non-zero exit the returned error is an
-// *exec.ExitError whose Stderr field carries the process's stderr, so callers
-// can still surface it in a failure message.
-func runSbxStdout(t *testing.T, ctx context.Context, args ...string) (string, error) {
-	t.Helper()
-	cmd := buildSbxCmd(ctx, args...)
-	out, err := cmd.Output()
-	return string(out), err
-}
-
-// shellQuote single-quotes a string for safe inclusion in an `sh -c` command.
-func shellQuote(s string) string {
-	return "'" + strings.ReplaceAll(s, "'", `'\''`) + "'"
+	tck.RunE2EKit(t, kitPath, tck.E2EOptions{AppName: appName})
 }
